Fix remove for Heap, comment within QueueLFCO implementation
Previous implementation only reheapDown the last value after putting it
at the position of the element to be delete, which may be incorrect for
some case (which the test did not catch). This is fixed by compare the
value to be deleted and the last element's value and choose reheapUp or
reheapDown occordingly.

QueueLFCO mentions 'total ordering', which does not mean what I think
that means. So fix to clarify meaning.