Finish wrapper for Huffman Coding tree (tested)
The tree's implementation is taken from the book. Destructor is written
so that those structure do not leak memory.

A wrapper class is implemented that generate the Huffman tree, freeing
necessary resources used in the process, traverse the finished tree to
extract the coding as a binary string and stores that coding into a
unordered_map for quick translation.