Change HuffmanCoding implementation to fit requirement (tested)
The new requirement states that for case that node have the same freq,
fallback to comparing ASCII value if the node is a leaf node, else
comparing the recency of its insertion into the heap.

The queue used to generate the Huffman tree is change to QueueLFCO
because of its ability to fallback to recency of insertion. New
constructor is implemented for QueueLFCO to generate a queue from
preexisting vector of element.

The previous Heap implementation (which is copied from the textbook) is
kept as is, because I don't want to remove the file and all its test
code.