Finish the system (tested)
I don't have much time left, and I'm tired. This is a very big change to
a lot of part within the project, and I don't really want to describe
them all. So to anyone reading this in the future trying to figure out
what the heck this commit do, I'm very sorry.